requirement.txt

requests
beautifulsoup4
gitpython

